# LEMAConsole/config/apiDoc.json
{
  "swagger": "2.0",
  "info": {
    "description": "API Documentation for LEMAConsole_EXT, LEMAConsole_INT, & LEMAgent. Project Summary: Lema aims to provide a cohesive and powerful solution to interacting with the Raspberry Pi's GPIO capabilities. The Lema Agent runs on the Raspberry Pi and communicates with the Lema Console, which provides a GUI webpage and API pathways for external integration.",
    "version": "1.0.0",
    "title": "Lema API",
    "contact": {
      "name": "Github Repository",
      "url": "https://github.com/RAK3RMAN/Lema"
    },
    "license": {
      "name": "MIT License",
      "url": "https://github.com/RAK3RMAN/Lema/blob/master/LICENSE"
    }
  },
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/lema-agent/broadcast": {
      "get": {
        "tags": [
          "LEMAgent"
        ],
        "summary": "Get information about node",
        "description": "",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/lema-agent/setup": {
      "post": {
        "tags": [
          "LEMAgent"
        ],
        "summary": "Request to setup node after discovery",
        "description": "",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Body parameters that are sent inside of api request",
            "required": true,
            "schema": {
              "$ref": "#/definitions/agent_setup"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
...
    "agent_setup": {
      "type": "object",
      "properties": {
        "console_ip": {
          "type": "string"
        },
        "console_secret": {
          "type": "string"
        }
      },
      "xml": {
        "name": "node"
      }
    }
  }
}